requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
langgraph==0.3.21
langchain-google-genai==2.1.2
langgraph-prebuilt==0.1.7
google-genai==1.7.0
python-dotenv